Comments router.js
RE line 125 - it seems like you're having issues because the database command is creating a new movie instead of updating a current movie on file. But the put route seems to be working, I see you've used method-override as well to get that put route to work (via Colt's Udemy course).